libsmm_acc: fix a concurrency bug
The static variable `cpu_fallback`, once set to `false` by one
thread, was causing other threads to fall back to the CPU although
they could launch a kernel on the GPU.

Solve this issue by determining whether to fall back to the CPU
using the state of the "kernel handle".

Also introduce a typedef and some simplifications for readability.